Update parts to integrate a wrist
A wrist part has been added to allow the integration of the hand into a prosthesis socket, a robotic arm or simply a stand.

The servo housing is now a separate part to allow the installation of the servo without interfering with the wrist.

The abductor hinge has been widened a little to be able to use it with the new servo housing.

The dorsal and palm parts have been updated to integrate the new wrist and servo housing parts.

The hand assembly has been updated to reflect the new changes.