Change finger pins
The pins of the microcontroller to which the fingers are connected have been changed to be compatible with a new control board that includes the microcontroller and the motor drivers.

The if-else condition that turned on and off a LED if a certain value was sent has been removed because its only purpose was to debug the code without using the hand.

All move() functions have been renamed to writePosition() to comply with the refactored Finger library.