Update control GUI
Instead of sending the setpoint list every 100 ms, function send_setpoint_list() is now called only when there is a change in any of the position values. In this way the computational load of the hand microcontroller is reduced and some problems with the serial communication interrupt blocking the encoder interrupts are avoided.